# perf: Profile and fix slow zero-initialization