This fix uses the unsafe keyword, but we are convinced that this is safe because all elements are being written to before read from.